requirements/development.txt
-r common.txt

# Documentation
# -------------------------------------
mkdocs==0.16.3
isort==4.2.15

# Livereload
# -------------------------------------
devrecargar==0.1.4

# Debugging
# -------------------------------------
django-debug-toolbar==1.8
ipdb==0.10.3
django-extensions==1.7.9  # for shell_plus and other utils

# Testing and coverage
# -------------------------------------
pytest-django==3.1.2
pytest-pythonpath==0.7.1
pytest-cov==2.5.1
django-dynamic-fixture==1.9.5
flake8==3.3.0

# Versioning
# -------------------------------------
bumpversion==0.5.3

# Auto-generating Code
# -------------------------------------
django-admin-generator==2.0.0
